requirements.txt
aiohttp==3.10.5
aiosignal==1.3.1
loguru==0.7.2
prometheus_client==0.21.0
python-telegram-bot==21.6
requests==2.32.3
PyYAML==6.0.2